+++ src/main/java/run/backend/domain/crew/service/CrewService.java
@@ -1,38 +1,117 @@
 package run.backend.domain.crew.service;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import run.backend.domain.crew.dto.common.CrewInviteCodeDto;
 import run.backend.domain.crew.dto.request.CrewInfoRequest;
 import run.backend.domain.crew.dto.response.*;
 import run.backend.domain.crew.entity.Crew;
+import run.backend.domain.crew.entity.JoinCrew;
+import run.backend.domain.crew.enums.JoinStatus;
+import run.backend.domain.crew.exception.CrewException;
+import run.backend.domain.crew.mapper.CrewMapper;
+import run.backend.domain.crew.repository.CrewRepository;
+import run.backend.domain.crew.repository.JoinCrewRepository;
+import run.backend.domain.file.service.FileService;
 import run.backend.domain.member.entity.Member;
 import run.backend.domain.member.enums.Role;
+import run.backend.domain.member.repository.MemberRepository;
 
-import java.time.YearMonth;
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class CrewService {
 
-public interface CrewService {
+    private final CrewMapper crewMapper;
+    private final FileService fileService;
+    private final CrewRepository crewRepository;
+    private final MemberRepository memberRepository;
+    private final JoinCrewRepository joinCrewRepository;
 
-    CrewInviteCodeDto createCrew(Member member, String imageStatus, MultipartFile image, CrewInfoRequest crewInfoRequest);
+    @Transactional
+    public CrewInviteCodeDto createCrew(Member member, String imageStatus, MultipartFile image, CrewInfoRequest data) {
 
-    void updateCrew(Member member, Crew crew, String imageStatus, MultipartFile image, CrewInfoRequest crewInfoRequest);
+        if (joinCrewRepository.existsByMemberAndJoinStatus(member, JoinStatus.APPROVED))
+            throw new CrewException.AlreadyJoinedCrew();
 
-    CrewInviteCodeDto getCrewInviteCode(Crew crew);
+        // 1. Crew 생성
+        String imageName = "default-profile-image.png";
+        if (imageStatus.equals("updated"))
+            imageName = fileService.saveProfileImage(image);
+        Crew crew = Crew.builder()
+                .image(imageName)
+                .name(data.name())
+                .description(data.description())
+                .build();
+        crewRepository.save(crew);
 
-    CrewProfileResponse getCrewByInviteCode(String inviteCode);
+        // 2. JoinCrew 생성
+        JoinCrew joinCrew = JoinCrew.createLeaderJoin(member, crew);
+        joinCrewRepository.save(joinCrew);
 
-    void joinCrew(Member member, Long crewId);
+        // 3. Member Role LEADER 로 변경
+        member.updateRole(Role.LEADER);
+        memberRepository.save(member);
 
-//    CrewSearchResponse searchCrew(String crewName);
-//
-//    CrewInfoResponse getCrewInfo(Long crewId);
-//
-//    CrewMonthlyCanlendarResponse getCrewMonthlyCalendar(Long crewId, YearMonth yearMonth);
-//
-//    CrewUpcomingEventResponse getUpcomingEvents(Long crewId);
-//
-//    CrewMemberResponse getCrewMemberProfile(Long crewId);
-//
-//    void updateCrewMemberRole(Long memberId, Role role);
-//
-//    CrewSearchResponse getRankCrew();
+        return new CrewInviteCodeDto(crew.getInviteCode());
+    }
+
+    @Transactional
+    public void updateCrew(Member member, Crew crew, String imageStatus, MultipartFile image, CrewInfoRequest data) {
+
+        switch (imageStatus) {
+
+            case "updated" :
+                fileService.deleteImage(crew.getImage());
+                String newImageName = fileService.saveProfileImage(image);
+                crew.updateImage(newImageName);
+                break;
+            case "removed" :
+                fileService.deleteImage(crew.getImage());
+                crew.updateImage("default-profile-image.png");
+                break;
+        }
+
+        if (data.name() != null)
+            crew.updateName(data.name());
+        if (data.description() != null)
+            crew.updateDescription(data.description());
+
+        crewRepository.save(crew);
+    }
+
+    public CrewInviteCodeDto getCrewInviteCode(Crew crew) {
+
+        return new CrewInviteCodeDto(crew.getInviteCode());
+    }
+
+    public CrewProfileResponse getCrewByInviteCode(String inviteCode) {
+
+        Crew crew = crewRepository.findByInviteCode(inviteCode)
+                .orElseThrow(CrewException.NotFoundCrew::new);
+        Member leader = joinCrewRepository.findCrewLeader(Role.LEADER, crew);
+
+        return crewMapper.toCrewProfile(crew, leader);
+    }
+
+    @Transactional
+    public void joinCrew(Member member, Long crewId) {
+
+        Crew crew = crewRepository.findById(crewId)
+                .orElseThrow(CrewException.NotFoundCrew::new);
+        if (joinCrewRepository.existsByMemberAndJoinStatus(member, JoinStatus.APPROVED))
+            throw new CrewException.AlreadyJoinedCrew();
+
+        JoinCrew joinCrew = JoinCrew.createAppliedJoin(member, crew);
+        joinCrewRepository.save(joinCrew);
+    }
+
+    public CrewBaseInfoResponse getCrewBaseInfo(Crew crew) {
+
+        int rank = 0;   // [TODO] : 스케줄링 rank 계산 구현 수정 예정
+
+        return crewMapper.toCrewBaseInfo(rank, crew);
+    }
 }
